+++ src/pages/DetailPost/style.js
@@ -1,7 +1,17 @@
 import styled from 'styled-components'
+
 export const Footer = styled.div`
 	position: absolute;
 	width: 100%;
 	left: 0;
 	bottom: 0;
+`
+
+export const CommentArea = styled.div`
+	width: 100%;
+	overflow: scroll;
+	height: 25.2rem;
+	::-webkit-scrollbar {
+		display: none;
+	}
 `
