+++ src/shared/hooks/usePosts.js
@@ -1,0 +1,57 @@
+import { useState } from 'react'
+import apis from 'shared/api'
+import { TEST_CHANNEL_ID } from 'shared/constants/chanelId'
+import { parsePostData, stringifyPostData } from 'shared/utils/postData'
+
+const usePosts = (user, timers) => {
+	const [posts, setPosts] = useState([])
+	const [isLoading, setIsLoading] = useState(false)
+
+	const getPosts = async () => {
+		setIsLoading(true)
+		const data = await apis.getPosts(TEST_CHANNEL_ID)
+
+		const filteredData = data.filter(post => {
+			if (post.title === 'Test') return false
+			const { share } = parsePostData(post.title)
+			return share === 'PUBLIC'
+		})
+
+		const fetchData = filteredData.map(post => {
+			const { timers } = parsePostData(post.title)
+			const like = post.likes.find(like => like.user === user._id)
+			const likeId = like ? like._id : null
+			return { ...post, timers, like, likeId, imageSrc: post.author.image }
+		})
+
+		setPosts(fetchData)
+		setIsLoading(false)
+	}
+
+	const updateMyPost = async () => {
+		const timersData = Object.keys(timers).map((key, idx) => {
+			return {
+				id: idx,
+				name: timers[key].name,
+			}
+		})
+		const data = stringifyPostData('PUBLIC', timersData)
+
+		await apis.modifyPost({
+			postId: user.posts[0]._id,
+			title: data,
+			image: null,
+			channelId: TEST_CHANNEL_ID,
+		})
+		getPosts()
+	}
+
+	const makePrivateMyPost = async () => {
+		await apis.disablePost(user.posts[0]._id, TEST_CHANNEL_ID)
+		getPosts()
+	}
+
+	return { posts, isLoading, getPosts, updateMyPost, makePrivateMyPost }
+}
+
+export default usePosts
